+++ scripts/test_yosys.sh
@@ -30,6 +30,9 @@ synthesize "${run_dir}/generic-switch.sv" main
 bash ngen.bat -fermat 4 -n 5 -k 3 -r 1 -pe 2 -architecture streamed -o "${run_dir}/fermat.sv" ntt
 synthesize "${run_dir}/fermat.sv" main
 
+bash ngen.bat -n 3 -rns-q 17,97 -rns-root 9,64 -rns-psi 3,8 -o "${run_dir}/rns.sv" -top RnsPolynomialMultiplier rnspolymul
+synthesize "${run_dir}/rns.sv" RnsPolynomialMultiplier
+
 bash ngen.bat -preset yata8 -k 3 -r 3 -o "${run_dir}/yata.sv" raintt
 synthesize "${run_dir}/yata.sv" SmallYata8RainttP27Rtl